requirements.txt
defusedxml==0.7.1
fonttools==4.53.1
fpdf2==2.7.9
numpy==2.1.1
pdf2image==1.17.0
pillow==10.4.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
six==1.16.0
wxPython==4.2.1
